Code related to storage class creation is modified in supervisour cluster tests (#1417)

+++ tests/e2e/multi_master_k8s.go
@@ -105,8 +105,10 @@ var _ = ginkgo.Describe("[csi-multi-master-block-e2e]", func() {
 		}
 
 		if sc != nil {
-			err = client.StorageV1().StorageClasses().Delete(ctx, sc.Name, *metav1.NewDeleteOptions(0))
-			gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			if !supervisorCluster {
+				err = client.StorageV1().StorageClasses().Delete(ctx, sc.Name, *metav1.NewDeleteOptions(0))
+				gomega.Expect(err).NotTo(gomega.HaveOccurred())
+			}
 		}
 
 		ginkgo.By("Waiting for old vsphere-csi-controller pod to be removed")
